FIX: Erros corrigidos #40

+++ core/models/avaliacao.py
@@ -7,7 +7,7 @@ from .produto import Produto
 class Avaliacao(models.Model):
    usuario = models.ForeignKey(User, on_delete=models.PROTECT, related_name="avaliacao", null=True, blank=True)
    produto = models.ForeignKey(Produto, on_delete=models.PROTECT, related_name="produto_avaliacao", null=True, blank=True)
-   nota = models.DecimalField( max_digits=3, decimal_places=2, null=False, blank=False)
+   nota = models.IntegerField(  null=False, blank=False)
    comentario = models.CharField(max_length=500, null=True, blank=True)
    data = models.DateTimeField(default=timezone.now)
